Ddded the process graph working with opendatacube
I'm not using merge_cubes with overlap resolver, instead I rename the bands before merging and use reduce_dimension with array_element after merge_cubes. I created this with VITO eb editor, but unfortunately if I copy and paste it into a new window it doesn't display it.
This graph still misses the last part with the MODIS data, which needs resample spatial and temporal.